Refactor TestDefinitions into own modules
