+++ src/cloudai/schema/test_template/chakra_replay/report_generation_strategy.py
@@ -25,7 +25,6 @@ from bokeh.models import ColumnDataSource, DataTable, Div, TableColumn, Title
 from bokeh.palettes import Turbo256
 from bokeh.plotting import figure, output_file, save
 from bokeh.transform import cumsum
-
 from cloudai import ReportGenerationStrategy
