+++ src/cloudai/schema/test_template/nccl_test/report_generation_strategy.py
@@ -19,7 +19,6 @@ from pathlib import Path
 from typing import List, Optional, Tuple
 
 import pandas as pd
-
 from cloudai import ReportGenerationStrategy
 from cloudai.report_generator.tool.bokeh_report_tool import BokehReportTool
 from cloudai.report_generator.tool.csv_report_tool import CSVReportTool
