+++ src/cloudai/systems/kubernetes/kubernetes_system.py
@@ -19,11 +19,10 @@ import time
 from pathlib import Path
 from typing import Any, Dict, List, Optional, Tuple, cast
 
-from kubernetes import client, config
-from kubernetes.client import ApiException, CustomObjectsApi, V1DeleteOptions, V1Job
-
 from cloudai import BaseJob, System
 from cloudai.runner.kubernetes.kubernetes_job import KubernetesJob
+from kubernetes import client, config
+from kubernetes.client import ApiException, CustomObjectsApi, V1DeleteOptions, V1Job
 
 
 class KubernetesSystem(System):
