+++ src/cloudai/test_definitions/gpt.py
@@ -1,0 +1,36 @@
+# SPDX-FileCopyrightText: NVIDIA CORPORATION & AFFILIATES
+# Copyright (c) 2024 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from .jax_toolbox import JaxFdl, JaxToolboxCmdArgs, JaxToolboxTestDefinition
+
+
+class GPTFdl(JaxFdl):
+    """GPT FDL configuration."""
+
+    checkpoint_policy: str = "save_nothing"
+
+
+class GPTCmdArgs(JaxToolboxCmdArgs):
+    """GPT JAX Toolbox test command arguments."""
+
+    fdl_config: str
+    fdl: GPTFdl
+
+
+class GPTTestDefinition(JaxToolboxTestDefinition):
+    """Test object for GPT."""
+
+    cmd_args: GPTCmdArgs
