+++ src/cloudai/test_definitions/nemo_launcher.py
@@ -14,119 +14,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import Literal, Optional
+from typing import Literal
 
-from cloudai import CmdArgs, TestDefinition
 from pydantic import BaseModel, ConfigDict, Field, field_validator
 
-
-class UCCCmdArgs(CmdArgs):
-    """UCC test command arguments."""
-
-    docker_image_url: str = Field(default="nvcr.io/nvidia/pytorch:24.02-py3")
-    collective: Literal[
-        "allgather",
-        "allgatherv",
-        "allreduce",
-        "alltoall",
-        "alltoallv",
-        "barrier",
-        "bcast",
-        "gather",
-        "gatherv",
-        "reduce",
-        "reduce_scatter",
-        "reduce_scatterv",
-        "scatter",
-        "scatterv",
-        "memcpy",
-        "reducedt",
-        "reducedt_strided",
-    ] = Field(default="alltoall")
-    b: int = Field(default=1)
-    e: str = Field(default="8M")
-
-
-class UCCTestDefinition(TestDefinition):
-    """Test object for UCC."""
-
-    cmd_args: UCCCmdArgs
-
-
-class NCCLCmdArgs(CmdArgs):
-    """NCCL test command arguments."""
-
-    docker_image_url: str = Field(default="nvcr.io/nvidia/pytorch:24.02-py3")
-    subtest_name: Literal[
-        "all_reduce_perf_mpi",
-        "all_gather_perf_mpi",
-        "alltoall_perf_mpi",
-        "broadcast_perf_mpi",
-        "gather_perf_mpi",
-        "hypercube_perf_mpi",
-        "reduce_perf_mpi",
-        "reduce_scatter_perf_mpi",
-        "scatter_perf_mpi",
-        "sendrecv_perf_mpi",
-        "bisection_perf_mpi",
-    ] = Field(default="all_reduce_perf_mpi")
-    nthreads: int = Field(default=1)
-    ngpus: int = Field(default=1)
-    minbytes: str = Field(default="32M")
-    maxbytes: str = Field(default="32M")
-    stepbytes: str = Field(default="1M")
-    op: Literal["sum", "prod", "min", "max", "avg", "all"] = Field(default="sum")
-    datatype: str = Field(default="float")
-    root: int = Field(default=0)
-    iters: int = Field(default=20)
-    warmup_iters: int = Field(default=5)
-    agg_iters: int = Field(default=1)
-    average: int = Field(default=1)
-    parallel_init: int = Field(default=0)
-    check: int = Field(default=1)
-    blocking: int = Field(default=0)
-    cudagraph: int = Field(default=0)
-
-
-class NCCLTestDefinition(TestDefinition):
-    """Test object for NCCL."""
-
-    cmd_args: NCCLCmdArgs
-
-    def extra_args_str(self) -> str:
-        parts = []
-        for k, v in self.extra_cmd_args.items():
-            parts.append(f"{k} {v}" if v else k)
-        return " ".join(parts)
-
-
-class ChakraReplayCmdArgs(CmdArgs):
-    """ChakraReplay test command arguments."""
-
-    docker_image_url: str = Field(default="DOCKER_IMAGE_URL")
-    mpi: str = Field(default="pmix")
-    trace_type: str = Field(default="et")
-    trace_path: Optional[str] = None
-    backend: str = Field(default="nccl")
-    device: str = Field(default="cuda")
-
-
-class ChakraReplayTestDefinition(TestDefinition):
-    """Test object for ChakraReplay."""
-
-    cmd_args: ChakraReplayCmdArgs
-
-
-class SleepCmdArgs(CmdArgs):
-    """Sleep test command arguments."""
-
-    seconds: int = Field(default=5)
-
-
-class SleepTestDefinition(TestDefinition):
-    """Test object for Sleep."""
-
-    cmd_args: SleepCmdArgs
+from cloudai import CmdArgs, TestDefinition
 
 
 class NumaMapping(BaseModel):
@@ -251,64 +143,3 @@ class NeMoLauncherTestDefinition(TestDefinition):
     """Test object for NeMoLauncher."""
 
     cmd_args: NeMoLauncherCmdArgs
-
-
-class JaxFdl(BaseModel):
-    """JAX FDL configuration."""
-
-    model_config = ConfigDict(extra="forbid")
-    num_gpus: int = 64
-    num_groups: int = 64
-    max_steps: int = 20
-    num_layers: int = 1
-    num_stages: int = 1
-    num_microbatches: int = 1
-    use_repeated_layer: bool = False
-    percore_batch_size: int = 1
-    checkpoint_policy: str = "save_nothing"
-    ici_mesh_shape: str = "'[1, 1, 8, 1]'"
-    dcn_mesh_shape: str = "'[1, 8, 1, 1]'"
-
-
-class JaxToolboxCmdArgs(CmdArgs):
-    """JAX Toolbox test command arguments."""
-
-    pass
-
-
-class JaxToolboxTestDefinition(TestDefinition):
-    """Test object for JAX Toolbox."""
-
-    pass
-
-
-class GrokCmdArgs(JaxToolboxCmdArgs):
-    """Grok test command arguments."""
-
-    fdl: JaxFdl
-    fdl_config: Optional[str] = None
-
-
-class GrokTestDefinition(JaxToolboxTestDefinition):
-    """Test object for Grok."""
-
-    cmd_args: GrokCmdArgs
-
-
-class GPTFdl(JaxFdl):
-    """GPT FDL configuration."""
-
-    checkpoint_policy: str = "save_nothing"
-
-
-class GPTJaxToolboxCmdArgs(JaxToolboxCmdArgs):
-    """GPT JAX Toolbox test command arguments."""
-
-    fdl_config: str
-    fdl: GPTFdl
-
-
-class GPTTestDefinition(JaxToolboxTestDefinition):
-    """Test object for GPT."""
-
-    cmd_args: GPTJaxToolboxCmdArgs
